feat: extract attack resolver
